+++ src/BaseApiCodes.php
@@ -98,21 +98,13 @@ class BaseApiCodes
         $tpl = 'response-builder::builder.http_%d';
 
         return [
-            /** @scrutinizer ignore-deprecated */
             self::OK()                          => 'response-builder::builder.ok',
-            /** @scrutinizer ignore-deprecated */
             self::NO_ERROR_MESSAGE()            => 'response-builder::builder.no_error_message',
-            /** @scrutinizer ignore-deprecated */
             self::EX_HTTP_EXCEPTION()           => 'response-builder::builder.http_exception',
-            /** @scrutinizer ignore-deprecated */
             self::EX_UNCAUGHT_EXCEPTION()       => 'response-builder::builder.uncaught_exception',
-            /** @scrutinizer ignore-deprecated */
             self::EX_HTTP_NOT_FOUND()           => sprintf($tpl, HttpResponse::HTTP_NOT_FOUND),
-            /** @scrutinizer ignore-deprecated */
             self::EX_HTTP_SERVICE_UNAVAILABLE() => sprintf($tpl, HttpResponse::HTTP_SERVICE_UNAVAILABLE),
-            /** @scrutinizer ignore-deprecated */
             self::EX_AUTHENTICATION_EXCEPTION() => sprintf($tpl, HttpResponse::HTTP_UNAUTHORIZED),
-            /** @scrutinizer ignore-deprecated */
             self::EX_VALIDATION_EXCEPTION()     => sprintf($tpl, HttpResponse::HTTP_UNPROCESSABLE_ENTITY),
         ];
     }
@@ -121,8 +113,6 @@ class BaseApiCodes
      * Returns API code for internal code OK
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function OK(): int
     {
@@ -143,8 +133,6 @@ class BaseApiCodes
      * Returns API code for internal code EX_HTTP_NOT_FOUND
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function EX_HTTP_NOT_FOUND(): int
     {
@@ -155,8 +143,6 @@ class BaseApiCodes
      * Returns API code for internal code EX_HTTP_EXCEPTION
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function EX_HTTP_EXCEPTION(): int
     {
@@ -167,8 +153,6 @@ class BaseApiCodes
      * Returns API code for internal code EX_UNCAUGHT_EXCEPTION
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function EX_UNCAUGHT_EXCEPTION(): int
     {
@@ -179,8 +163,6 @@ class BaseApiCodes
      * Returns API code for internal code EX_AUTHENTICATION_EXCEPTION
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function EX_AUTHENTICATION_EXCEPTION(): int
     {
@@ -191,8 +173,6 @@ class BaseApiCodes
      * Returns API code for internal code EX_VALIDATION_EXCEPTION
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function EX_VALIDATION_EXCEPTION(): int
     {
@@ -203,8 +183,6 @@ class BaseApiCodes
      * Returns API code for internal code EX_HTTP_SERVICE_UNAVAILABLE
      *
      * @return int valid API code in current range
-     *
-     * @deprecated Configure Exception Handler to use your own API code. This method will be removed in v8.
      */
     public static function EX_HTTP_SERVICE_UNAVAILABLE(): int
     {