+++ src/Converters/JsonSerializableConverter.php
@@ -17,7 +17,7 @@ namespace MarcinOrlowski\ResponseBuilder\Converters;
 use MarcinOrlowski\ResponseBuilder\Contracts\ConverterContract;
 use MarcinOrlowski\ResponseBuilder\Validator;
 
-class JsonSerializableConverter implements ConverterContract
+final class JsonSerializableConverter implements ConverterContract
 {
 	/**
 	 * Returns array representation of the object implementing \JsonSerializable interface.
@@ -32,6 +32,6 @@ class JsonSerializableConverter implements ConverterContract
 	{
 		Validator::assertInstanceOf('obj', $obj, \JsonSerializable::class);
 
-		return ['val' => json_decode(json_encode($obj->jsonSerialize()), true)];
+		return ['val' => \json_decode(json_encode($obj->jsonSerialize()), true)];
 	}
 }