+++ src/ResponseBuilderServiceProvider.php
@@ -70,7 +70,7 @@ class ResponseBuilderServiceProvider extends ServiceProvider
 
         $merged_config = Util::mergeConfig($defaults, $config);
 
-        if (array_key_exists('converter', $merged_config)) {
+        if (\array_key_exists('converter', $merged_config)) {
             Util::sortArrayByPri($merged_config['converter']);
         }