+++ tests/tests/Localization/TranslationTest.php
@@ -13,10 +13,7 @@ namespace MarcinOrlowski\ResponseBuilder\Tests;
  * @link      https://github.com/MarcinOrlowski/laravel-api-response-builder
  */
 
-use MarcinOrlowski\ResponseBuilder\Converter;
-use MarcinOrlowski\ResponseBuilder\ResponseBuilder;
 use Illuminate\Support\Facades\Config;
-use Symfony\Component\HttpFoundation\Response as HttpResponse;
 
 class TranslationTest extends TestCase
 {
@@ -46,7 +43,7 @@ class TranslationTest extends TestCase
                 }
             );
 
-        $this->assertGreaterThan(0, count($supported_languages));
+        $this->assertGreaterThan(0, \count($supported_languages));
 
         foreach ($supported_languages as $lang) {
             // get the translation array for given language